chore: move `util::{... build funcs ...}` -> `cargo_near_build::cargo_native::compile::run`
