+++ cargo-near-build/Cargo.toml
@@ -12,3 +12,4 @@ log = "0.4"
 bs58 = "0.5"
 hex = "0.4.3"
 sha2 = "0.10"
+cargo_metadata = "0.18.1"
