+++ cargo-near-build/src/cargo_native.rs
@@ -1,4 +1,6 @@
 // TODO: make module non-pub
+pub mod compile;
+// TODO: make module non-pub
 pub mod target;
 
 pub trait ArtifactType {
