+++ cargo-near/src/commands/build_command/build.rs
@@ -1,5 +1,5 @@
 use camino::Utf8PathBuf;
-use cargo_near_build::cargo_native::WASM;
+use cargo_near_build::cargo_native::{self, WASM};
 use cargo_near_build::pretty_print;
 use cargo_near_build::types::cargo::manifest_path::{ManifestPath, MANIFEST_FILE_NAME};
 use cargo_near_build::types::near::VersionMismatch;
@@ -12,7 +12,6 @@ use crate::commands::build_command::{
     NEP330_BUILD_COMMAND_ENV_KEY, NEP330_CONTRACT_PATH_ENV_KEY, NEP330_SOURCE_CODE_SNAPSHOT_ENV_KEY,
 };
 use crate::types::metadata::CrateMetadata;
-use crate::util;
 use crate::BuildArtifact;
 use cargo_near_build::types::color_preference::ColorPreference;
 
@@ -209,7 +208,7 @@ pub fn run(args: Opts) -> color_eyre::eyre::Result<BuildArtifact<WASM>> {
     }
 
     pretty_print::step("Building contract");
-    let mut wasm_artifact = util::compile_project::<WASM>(
+    let mut wasm_artifact = cargo_native::compile::run::<WASM>(
         &crate_metadata.manifest_path,
         &cargo_args,
         build_env,
